requirements.txt
streamlit
pandas
plotly
google-generativeai
python-dotenv
scipy
openpyxl
